Изменение main.js.  Разделение на модули, для пошаговой работы
