+++ js/main.js
@@ -1,58 +1,6 @@
-// Генерация случайного числа в диапазоне
-const getRandomInteger = (min, max) => {
-  return Math.floor(Math.random() * (max - min + 1)) + min;
-};
+import { generatePhotos } from './photo.js';
 
-// Создание генератора уникальных ID
-const createIdGenerator = () => {
-  let lastId = 0;
-  return () => ++lastId;
-};
-
-const generatePhotoId = createIdGenerator();
-const generateCommentId = createIdGenerator();
-
-// Данные для комментариев
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают.'
-];
-
-const NAMES = ['Артём', 'Мария', 'Дмитрий', 'Анна', 'Сергей', 'Елена', 'Алексей', 'Ольга', 'Иван', 'Наталья'];
-
-// Получение случайного элемента массива
-const getRandomArrayElement = (array) => array[getRandomInteger(0, array.length - 1)];
-
-// Создание комментария
-const createComment = () => ({
-  id: generateCommentId(),
-  avatar: `img/avatar-${getRandomInteger(1, 6)}.svg`,
-  message: getRandomArrayElement(MESSAGES),
-  name: getRandomArrayElement(NAMES)
-});
-
-// Создание фотографии
-const createPhoto = () => {
-  const photoId = generatePhotoId();
-
-  return {
-    id: photoId,
-    url: `photos/${photoId}.jpg`,
-    description: 'Красивый момент из жизни',
-    likes: getRandomInteger(15, 200),
-    comments: Array.from({length: getRandomInteger(0, 30)}, createComment)
-  };
-};
-
-// Генерация массива фотографий
-const generatePhotos = () => Array.from({length: 25}, createPhoto);
-
-// Создаем массив фотографий
 const photos = generatePhotos();
+console.log('Сгенерированные фотографии:', photos);
 
-// Для отладки выводим в консоль
-console.log(photos);
+export { photos };
